10.12.4

+++ Security/OSX/libsecurity_utilities/lib/pcsc++.cpp
@@ -190,7 +190,7 @@ void Session::close()
 
 bool Session::check(long rc)
 {
-	switch (rc) {
+	switch ((unsigned long) rc) {
 	case SCARD_S_SUCCESS:
 		return true;	// got reader(s), call succeeded
 	case SCARD_E_READER_UNAVAILABLE:
@@ -214,7 +214,7 @@ void Session::listReaders(vector<string> &readers, const char *groups)
 				decode(readers, mReaderBuffer, size);
 				return;
 			}
-		case SCARD_E_INSUFFICIENT_BUFFER:
+		case (int32_t) SCARD_E_INSUFFICIENT_BUFFER:
 			mReaderBuffer.resize(size);
 			break;
 		default:
